+++ src/main.js
@@ -3,12 +3,10 @@ import App from './App'
 
 import AvueFormDesign from '../packages/';
 import AvueUeditor from 'avue-plugin-ueditor'
-import AvueMap from 'avue-plugin-map'
 
 Vue.use(window.AVUE)
 Vue.use(AvueFormDesign)
 Vue.use(AvueUeditor)
-Vue.use(AvueMap)
 
 Vue.config.productionTip = false