+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## 更新日志
 
+*2020-07-06*
+
+#### 更新 [avue 2.6.10 版本](https://avuejs.com/doc/changelog) 
+
+- color组件 -> input-color。
+- icon-select组件 -> input-icon。
+- 删除avue-plugin-map依赖，使用input-map。
+
 *2020-06-29*
 
 #### 更新 [avue 2.6.8 版本](https://avuejs.com/doc/changelog)